refactor(everything): add some organisation to the codebase
breaking: this breaks everything. make sure to pull before contributing
