+++ src/pages/App.tsx
@@ -3,11 +3,11 @@ import { HeaderSection } from '../sections/Header';
 import { HeroSection } from '../sections/Hero';
 import { StatsSection } from '../sections/Statistics';
 import { FooterSection } from '../sections/Footer';
-import './style/App.css';
+import "../styles/App.css"
 
 function App() {
   return (
-    
+
     <div className="App">
       <header className="App-header">
         <HeaderSection />
